+++ utils/common.js
@@ -1,3 +1,4 @@
+import { NO_FISHING_WORLDS } from '../constants/areas';
 import { RED, DARK_GRAY, BLUE, WHITE, BOLD, RESET } from '../constants/formatting';
 import { NBTTagString } from '../constants/javaTypes';
 import { EntityFishHook, NBTTagString } from '../constants/javaTypes';
@@ -440,6 +441,35 @@ export function getMessageId() {
 export function getZoneName() {
 	const zoneLine = Scoreboard.getLines().find((line) => line.getName().includes('⏣'));
 	return zoneLine?.trim() || '';
+}
+
+/**
+ * Check whether current world name supports fishing features.
+ * @returns {boolean}
+ */
+export function isInFishingWorld(worldName) {
+	if (!worldName) return false;
+	return !NO_FISHING_WORLDS.includes(worldName);
+}
+
+/**
+ * Get MC entity ID.
+ * @param {Entity} entity
+ * @returns {number} ID
+ */
+export function getMcEntityId(entity) {
+    if (!entity || !entity.entity) return;
+    return entity.entity.func_145782_y(); // func_145782_y() -> getEntityId()
+}
+
+/**
+ * Get MC entity ID.
+ * @param {number} id MC entity ID
+ * @returns {object} MC entity
+ */
+export function getMcEntityById(id) {
+    if (!id) return;
+    return World.getWorld()?.func_73045_a(id); // func_73045_a() -> getEntityByID()
 }
 
 function getCurrentGuiChestName() {