+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ Other:
 - Refactored Rare catches tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Abandoned Quarry tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Jerry Workshop tracker to not execute any checks when setting is disabled / when being in a wrong world.
+- Refactored Totem tracker to not execute any checks when setting is disabled / when being in a wrong world.
 
 ## v1.36.0