+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ Bugfixes:
 - 
 
 Other:
+- Refactored Legion & Bobbing Time tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Rare catches tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Abandoned Quarry tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Jerry Workshop tracker to not execute any checks when setting is disabled / when being in a wrong world.