+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ Bugfixes:
 - Fixed Totem of Corruption expiration alert triggering on other player's totem sometimes.
 
 Other:
+- Refactored Legion & Bobbing Time tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Rare catches tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Abandoned Quarry tracker to not execute any checks when setting is disabled / when being in a wrong world.
 - Refactored Jerry Workshop tracker to not execute any checks when setting is disabled / when being in a wrong world.